+++ lib/mino/tasks/builtin.clj
@@ -61,7 +61,7 @@
    "src/prim/stateful.c" "src/prim/stm.c" "src/prim/agent.c" "src/prim/module.c"
    "src/prim/ns.c"
    "src/prim/fs.c" "src/prim/proc.c"
-   "src/prim/host.c" "src/interop/syntax.c"
+   "src/prim/host.c" "src/prim/jvm_statics.c" "src/interop/syntax.c"
    "src/collections/clone.c" "src/regex/re.c" "src/collections/transient.c"
    "src/async/scheduler.c" "src/async/timer.c" "src/async/chan.c"
    "src/prim/async.c"